Recreate a hung RabbitMQ container in wait_rabbitmq_to_start
The RabbitMQ container occasionally hangs on startup in CI: the entrypoint
produces no output at all, the Erlang node never registers with epmd, and
`rabbitmqctl await_startup` keeps failing for the whole 120 seconds - while
a parallel worker's container on the same host starts in 3 seconds. When
that happens, every test in the module fails with `Cannot wait RabbitMQ
container`. This has been striking roughly weekly since at least 2026-06-16,
almost always in `arm_binary` integration jobs.

Make `wait_rabbitmq_to_start` collect diagnostics, recreate the container
and wait once more before giving up. Also, `rabbitmq_debuginfo` used to
discard all of its output (piped to stdout and never read); route it through
`run_and_check` so it lands in the test log, and add the container state
and the in-container process list so a future hang shows where the
entrypoint is stuck.

Report: https://s3.amazonaws.com/clickhouse-test-reports/json.html?REF=master&sha=e4d3698282346fcd7e496ef668ff517106c84ac5&name_0=MasterCI&name_1=Integration%20tests%20%28arm_binary%2C%20distributed%20plan%2C%201%2F4%29&name_1=Integration%20tests%20%28arm_binary%2C%20distributed%20plan%2C%201%2F4%29
